changes to main.py UI - now takes the model name as the sole command line arg - has a looping prompt for new audio files to classify - searches for a directory ending in ".tf" if no model cmd arg is given

+++ main.py
@@ -1,13 +1,11 @@
 """
-Usage: python main.py <pathtoaudio> [<pathtomodel>]
+Usage: python main.py <pathtomodel>
 """
 
 
 import os
-os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3'
-import random
+os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3' # disable tensorflow warning messages about not having a cuda gpu
 import sys
-import io
 
 import tensorflow as tf
 from tensorflow import keras
@@ -29,44 +27,55 @@ if __name__ == '__main__':
 
 
     # load model
-    model_path = "./model.tf"
-    if len(sys.argv) > 2:
-        model_path = sys.argv[2]
+    model_path = ""
+    # if user has provided model path, use that
+    if len(sys.argv) > 1:
+        model_path = sys.argv[1]
+    # otherwise, search for directory ending in .tf
+    else: 
+        for item in os.listdir():
+            if item.endswith(".tf") and os.path.isdir(item):
+                model_path = item
+                break
 
     if os.path.exists(model_path):
+        print(f"Loading model from {model_path}...")
         model = keras.models.load_model(model_path)
+        print("Model loaded. Type 'exit' to exit.")
     else:
-        print(f"Model directory \"{model_path}\" not found.")
+        print(f"No model found.")
         exit(1)
         
-    # Verify audio
-    audio_path = ""
+    while True:
+        # get audio file from user
+        audio_path = input("Classify audio file: ")
+        if audio_path in ["exit", "stop", "quit"]:
+            break
+        if audio_path == "":
+            continue
 
-    if os.path.exists(sys.argv[1]):
-        audio_path = sys.argv[1]
-    else:
-        print("Audio file not found.")
-        exit(1)
+        try:
+            # create spectrogram
+            spec = convert_audio_to_spectogram(audio_path, normalize=True, augment=False)
+            spec.save(spec_temp_file_name)
 
-    # create spectrogram
-    spec = convert_audio_to_spectogram(audio_path, normalize=True, augment=False)
-    spec.save(spec_temp_file_name)
+            # preprocess the spectrogram
+            image = tf.keras.preprocessing.image.load_img(spec_temp_file_name, target_size=(100, 300))
+            input_arr = tf.keras.preprocessing.image.img_to_array(image)
+            input_arr = np.array([input_arr])
 
-    # preprocess the spectrogram
-    image = tf.keras.preprocessing.image.load_img(spec_temp_file_name, target_size=(100, 300))
-    input_arr = tf.keras.preprocessing.image.img_to_array(image)
-    input_arr = np.array([input_arr])
+            predictions = model.predict(input_arr)
 
-    predictions = model.predict(input_arr)
-    print(f'{predictions=}')
-    # print(f'{model.classnames}')
+            if np.argmax(predictions[0]) == 1:
+                print(f'Yes ({(100 * predictions[0][1]):.2f}%)')
+            elif np.argmax(predictions[0]) == 0:
+                print(f'No ({(100 * predictions[0][0]):.2f}%)')
+            else:
+                print("an error has occurred")
+            
+            # print(f'\t{predictions=}')
 
-    if np.argmax(predictions[0]) == 1:
-        print(f'Yes with confidence {2 * (predictions[0][1] - 0.5)}')
-    elif np.argmax(predictions[0]) == 0:
-        print(f'No with confidence {2 * (predictions[0][0] - 0.5)}')
-    else:
-        print("an error has occurred")
-
-    # clean up
-    os.remove(spec_temp_file_name)
+            # clean up
+            os.remove(spec_temp_file_name)
+        except Exception as e:
+            print(e)
